requirements.txt
pandas==2.1.4
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
